requirements.txt
asgiref==3.5.1
click==8.1.3
Django==4.0.4
django-cors-headers==3.11.0
django-filebrowser==4.0.1
django-grappelli==3.0.3
django-tinymce==3.4.0
djangorestframework==3.13.1
filebrowser==1.0.7
filebrowser-safe==1.1.1
Flask==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.1.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
Werkzeug==2.1.2
